+++ tests/Controller/Address/AddressDeleteTest.php
@@ -85,5 +85,4 @@ class AddressDeleteTest extends \PHPUnit_Framework_TestCase
         $out = json_decode((string) $response->getBody(), true);
         $this->assertEquals(2625, $out['status_code']);
     }
-
 }